Modernize the ixo address generation function
Apparently the "bytes" value for the crypto's digest function's
"encoding" argument is deprecated. It looks like it was used to
make the function output a Buffer data type, but the digest
function is currently behaving the same way even if we don't
specify an encoding.

This is no big deal in Node environments as it seems like this
legacy argument is still supported, but it breaks when we use the
React Native shim for crypto. One can say that it is susceptible
of breaking in similar scenarios.

Removing this parameter seems to be safe, in my experiments it
worked in both Node and the shimmed RN environments.

References:
https://github.com/mvayngrib/react-native-crypto